+++ tomcat/src/main/java/com/techcourse/service/Service.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.techcourse.service;
 
 import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
